Add 2.1.0 version of CodeAnalysis/CodeAnalysis.CSharp to prebuilt packages
We need real assets for Microsoft.CodeAnalysis.CSharp.dll in order for
the Microsoft.DotNet.ProjectJsonMigration tool to work.

Another option in the future would be to use the live version of Roslyn
CodeAnalysis packages but in order to match what shipped with the 2.1.4
SDK we are going to use a prebuilt package instead.